third_party/blink/web_tests/platform/generic/fast/js/codegen-temporaries-multiple-global-blocks-expected.txt

Tests whether bytecode codegen properly handles temporaries across multiple global blocks.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS v1 is 2
PASS o2.a is 2
PASS o3.a is 2
PASS r4 is 1
PASS o5.a is 2
PASS o6.a is 2
PASS o7.a is 2
PASS o8.a is 2
PASS successfullyParsed is true

TEST COMPLETE
